// app/types/theme.ts
export interface MyTheme {
  primaryColor: string;
  linkColor: string;
  successColor: string;
  warningColor: string;
  errorColor: string;
  fontSizeBase: string;
  headingColor: string;
  textColor: string;
  textColorSecondary: string;
  white: string;
  black: string;
  darkBlue: string;
  warning: string;
  secondaryColor: string;
  disabledColor: string;
  borderRadiusBase: string;
  borderColorBase: string;
  boxShadowBase: string;
  fontFamily: string;
  letterSpacing: string;
  grayBackground: string;
  darkGrayBackground: string;
  primary: string;
  secondary: string;
  divnotesGreen: string;
  divnotesPurple: string;
  disabled: string;
  selected: string;
  inchworm: string;
  paleCyan: string;
  peachOrange: string;
  pristine: string;
  lightGray: string;
  menu: string;
  background: string;
  add: string;
  alert: string;
  error: string;
  sales: string;
  success: string;
}